pyproject.toml
[project]
name = "mem0-mcp"
version = "0.1.0"
description = "MCP server for integrating long term memory into AI agents with Mem0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "fastmcp>=0.1.0",
    "mem0ai>=0.1.88",
    "vecs>=0.4.5",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "asyncio",
    "typing-extensions"
]
